feat: move DiscoverNIP89FeedFilter to commons
- Core filter/feed logic in commons with IAccount + ICacheProvider
- Android wrapper overrides sort() for participant-count ranking
